reorganize lang/syn/src/idl module
I've put the types into it's own module and everything related to
the file parsing into it's own module.
