+++ cli/src/config.rs
@@ -1,7 +1,7 @@
 use crate::is_hidden;
 use anchor_client::Cluster;
-use anchor_syn::idl::Idl;
-use anyhow::{anyhow, bail, Context, Error, Result};
+use anchor_syn::idl::types::Idl;
+use anyhow::{anyhow, Context, Error, Result};
 use clap::{Parser, ValueEnum};
 use heck::ToSnakeCase;
 use reqwest::Url;
